% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{bb_distribution}
\alias{bb_distribution}
\title{Beta-binomial distribution on the full support 0:n}
\usage{
bb_distribution(a, b, n)
}
\arguments{
\item{a}{Alpha parameter (> 0).}

\item{b}{Beta parameter (> 0).}

\item{n}{Number of trials / maximum score.}
}
\value{
A list with elements x (support 0:n), Px (pmf), cum (cdf) and
  perc (mid-p percentiles, i.e. P(X < x) + 0.5 * P(X = x)).
}
\description{
Computes the (normalized) probability mass function, the cumulative
distribution function and mid-p percentiles of a beta-binomial
distribution in a numerically stable way (log space).
}
\keyword{internal}
