% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildCnormObject}
\alias{buildCnormObject}
\title{Build cnorm object from data and bestModle model object}
\usage{
buildCnormObject(data, model)
}
\arguments{
\item{data}{A data object from 'prepareData', or from 'rankByGroup' and
'computePower'}

\item{model}{Object obtained from the bestModel function}
}
\value{
A cnorm object
}
\description{
Helper function to build a cnorm object from a data object and
a model object from the bestModel function for compatibility reasons.
}
\examples{
\dontrun{
  data <- prepareData(elfe)
  model <- bestModel(data, k = 4)
  model.cnorm <- buildCnormObject(data, model)
}

}
