% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{calcPolyInL}
\alias{calcPolyInL}
\title{Internal function for retrieving regression function coefficients at specific age}
\usage{
calcPolyInL(raw, age, model)
}
\arguments{
\item{raw}{The raw value (substracted from the intercept)}

\item{age}{The age}

\item{model}{The cNORM regression model}
}
\value{
The coefficients
}
\description{
The function is an inline for searching zeros in the inverse regression
function. It collapses the regression function at a specific age and simplifies
the coefficients.
}
