% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{calcPolyInLBase2}
\alias{calcPolyInLBase2}
\title{Internal function for retrieving regression function coefficients at specific
age (optimized)}
\usage{
calcPolyInLBase2(raw, age, coeff, k)
}
\arguments{
\item{raw}{The raw value (subtracted from the intercept)}

\item{age}{The age}

\item{coeff}{The cNORM regression model coefficients}

\item{k}{The cNORM regression model power parameter}
}
\value{
The coefficients
}
\description{
The function is an inline for searching zeros in the inverse regression
function. It collapses the regression function at a specific age and
simplifies the coefficients. Optimized version of the prior 'calcPolyInLBase'
}
