% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{check.cnorm}
\alias{check.cnorm}
\title{S3 method for checking the consistency of a cnorm regression model}
\usage{
check.cnorm(object, ...)
}
\description{
While abilities increase and decline over age, within one age group, the
norm scores always have to show a linear increase or decrease with increasing raw
scores. Violations of this assumption are a strong indication for problems
in modeling the relationship between raw and norm scores.
}
\seealso{
checkConsistency
}
