% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking.R
\name{checkWeights}
\alias{checkWeights}
\title{Check, if NA or values <= 0 occur and issue warning}
\usage{
checkWeights(weights)
}
\arguments{
\item{weights}{Raking weights}
}
\description{
Check, if NA or values <= 0 occur and issue warning
}
\keyword{internal}
