% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNORM.R
\name{cNORM.GUI2}
\alias{cNORM.GUI2}
\title{Launch the cNORM Parametric Modeling Shiny Application}
\usage{
cNORM.GUI2(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical, whether to launch browser automatically (default: TRUE)}
}
\description{
Launch the cNORM Parametric Modeling Shiny Application
}
