% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{cnorm.lasso}
\alias{cnorm.lasso}
\title{Fit a LASSO Regression Model for Continuous Norming with Standardization}
\usage{
cnorm.lasso(
  raw,
  age,
  weights = NULL,
  scale = "T",
  descend = FALSE,
  k = 5,
  t = 3,
  nfolds = 10,
  max_attempts = 100,
  plot = TRUE
)
}
\arguments{
\item{raw}{A numeric vector of raw scores.}

\item{age}{A numeric vector of age values corresponding to the raw scores.}

\item{weights}{An optional numeric vector of weights for the observations.}

\item{scale}{type of norm scale, either T (default), IQ, z or percentile (= no
transformation); a double vector with the mean and standard deviation can as
well, be provided f. e. c(10, 3) for Wechsler scale index points}

\item{descend}{ranking order (default descent = FALSE): inverses the
ranking order with higher raw scores getting lower norm scores; relevant
for example when norming error scores, where lower scores mean higher
performance}

\item{k}{The maximum power for location terms in the model (default is 5).}

\item{t}{The maximum power for age terms in the model (default is 3).}

\item{nfolds}{The number of folds to use in cross-validation (default is 10).}

\item{max_attempts}{The maximum number of lambda values to try when searching for a monotonic model (default is 100).}

\item{plot}{Logical; if TRUE, plots the fitted model (default is FALSE).}
}
\value{
An object of class "cnormLasso" containing the fitted model and various model details.
}
\description{
This function fits a LASSO regression model for continuous norming, with options for
ensuring monotonicity and cross-validation. It standardizes age, location, and raw scores
before fitting the model.
}
