% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{cv.cnorm}
\alias{cv.cnorm}
\title{S3 method for cross validation of cnorm objects}
\usage{
cv.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object or ranked data}

\item{...}{Additional parameters}
}
\description{
This function helps in selecting the number of terms for the model by doing repeated
Monte Carlo cross validation with 80 percent of the data as training data and 20 percent as
the validation data. The cases are drawn randomly but stratified by norm group. Successive
models are retrieved with increasing number of terms and the RMSE of raw scores (fitted by
the regression model) is plotted for the training, validation and the complete dataset.
Additionally to this analysis on the raw score level, it is possible (default) to estimate
the mean norm score reliability and crossfit measures.
Please apply to the data (e. g. cv(model$data)) for model selection and to model itself
(cv(model)) to estimate precision of predefined regression function.
}
\seealso{
cnorm.cv
}
