% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{derivationPlot}
\alias{derivationPlot}
\title{Plot first order derivation of regression model}
\usage{
derivationPlot(model, minAge = 2, maxAge = 5, minNorm = 25,
  maxNorm = 75, stepAge = 0.2, stepNorm = 1)
}
\arguments{
\item{model}{The model from the bestModel function}

\item{minAge}{Age to start with checking}

\item{maxAge}{Upper end of the age check}

\item{minNorm}{Lower end of the norm value range, in case of T values, 25 might be good}

\item{maxNorm}{Upper end of the norm value range, in case of T values, 25 might be good}

\item{stepAge}{Stepping parameter for the age check, usually 1 or 0.1; lower
values indicate higher precision / closer checks}

\item{stepNorm}{Stepping parameter for norm values}
}
\description{
Plots the values obtained via the first derivation of the regression model
in dependence of the norm value. The results indicate the progression of the
norm values within each age group. The regression based modeling approach
relies on the assumption of a linear progression of the norm values.
Negative values in the first order derivation indicate a violation of this
assumption. Values near zero
are typical for bottom and ceiling effects in the raw data.
The regression models usually converge within the range of the original
values. In case of vertical and horizontal extrapolation, with increasing
distance to the original data, the risk of assumption violation increases
as well.
}
\examples{
# Load example data set, compute model and plot results
normData <- prepareData()
m <- bestModel(data = normData)
derivationPlot(m, minAge=2, maxAge=5, step=.2, minNorm=25, maxNorm=75, stepNorm=1)
}
