% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{diagnostics}
\alias{diagnostics}
\title{Model diagnostics for parametric models}
\usage{
diagnostics(model, ...)
}
\arguments{
\item{model}{A cnormBetaBinomial, cnormShash or cnormCMP object}

\item{...}{Further arguments (e.g. age, score, weights) passed on}
}
\description{
Model diagnostics for parametric models
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=modelSummary]{modelSummary()}},
\code{\link[=normTable]{normTable()}},
\code{\link[=printSubset]{printSubset()}},
\code{\link[=rangeCheck]{rangeCheck()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=weightedAverageModel]{weightedAverageModel()}}
}
\concept{model}
