% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{dshash}
\alias{dshash}
\title{Sinh-Arcsinh (shash) distribution functions}
\usage{
dshash(x, mu = 0, sigma = 1, epsilon = 0, delta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter (> 0)}

\item{epsilon}{skewness parameter}

\item{delta}{tail weight parameter (> 0)}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Density, distribution function, quantile function and random generation
for the Sinh-Arcsinh distribution
}
\keyword{internal}
