% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{logLik.cnormCMP}
\alias{logLik.cnormCMP}
\title{Log-likelihood of a CMP model (enables \code{AIC()} and \code{BIC()})}
\usage{
\method{logLik}{cnormCMP}(object, ...)
}
\arguments{
\item{object}{A fitted model of class "cnormCMP".}

\item{...}{Not used.}
}
\value{
An object of class "logLik".
}
\description{
Log-likelihood of a CMP model (enables \code{AIC()} and \code{BIC()})
}
