% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{nobs.cnormCMP}
\alias{nobs.cnormCMP}
\title{Number of observations used to fit a CMP model}
\usage{
\method{nobs}{cnormCMP}(object, ...)
}
\arguments{
\item{object}{A fitted model of class "cnormCMP".}

\item{...}{Not used.}
}
\value{
Integer.
}
\description{
Number of observations used to fit a CMP model
}
