% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.cnorm}
\alias{plot.cnorm}
\title{S3 function for plotting cnorm objects}
\usage{
\method{plot}{cnorm}(x, y, ...)
}
\arguments{
\item{x}{the cnorm object}

\item{y}{the type of plot as a string, can be one of
'raw' (1), 'norm' (2), 'curves' (3), 'percentiles' (4), 'series' (5), 'subset' (6),
or 'derivative' (7), either as a string or the according index}

\item{...}{additional parameters for the specific plotting function}
}
\description{
S3 function for plotting cnorm objects
}
\seealso{
Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
