% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{plot.cnormShash}
\alias{plot.cnormShash}
\title{Plot Sinh-Arcsinh Model with Data and Percentile Lines}
\usage{
\method{plot}{cnormShash}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "cnormShash"}

\item{...}{Additional arguments including age, score, weights, percentiles, points}
}
\value{
A ggplot object
}
\description{
Plot Sinh-Arcsinh Model with Data and Percentile Lines
}
