% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.curves.cnorm}
\alias{plot.curves.cnorm}
\title{S3 method for plotting percentile curves of cnorm objects}
\usage{
plot.curves.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the norm curves based on the regression model.
Please check the function for inconsistent curves: The different
curves should not intersect. Violations of this assumption are a strong
indication for violations of model assumptions in modeling the relationship between raw
and norm scores.
}
\seealso{
plotNormCurves

Other plot: 
\code{\link{plot.cnorm}()},
\code{\link{plot.density.cnorm}()},
\code{\link{plot.derivative.cnorm}()},
\code{\link{plot.norm.cnorm}()},
\code{\link{plot.raw.cnorm}()},
\code{\link{plot.series.cnorm}()},
\code{\link{plot.subset.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNormCurves}()},
\code{\link{plotNorm}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
