% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.derivative.cnorm}
\alias{plot.derivative.cnorm}
\title{S3 method for plotting derivatieves of the regression function of cnorm objects}
\usage{
plot.derivative.cnorm(object, ...)
}
\description{
Plots the scores obtained via the first order derivative of the regression model
in dependence of the norm score. The results indicate the progression of the
norm scores within each age group. The regression based modeling approach
relies on the assumption of a linear progression of the norm scores.
Negative scores in the first order derivative indicate a violation of this
assumption. Scores near zero are typical for bottom and ceiling effects in the raw data.
}
\seealso{
plotDerivative
}
