% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.norm.cnorm}
\alias{plot.norm.cnorm}
\title{S3 method for plotting manifest against predicted norm sores
The function plots the manifest norm score against the fitted norm score from
the inverse regression model per group. This helps to inspect the precision
of the modeling process. The scores should not deviate too far from
regression line. The computation of the standard error is based on Oosterhuis, van der
Ark and Sijtsma (2016).}
\usage{
plot.norm.cnorm(object, ...)
}
\description{
S3 method for plotting manifest against predicted norm sores
The function plots the manifest norm score against the fitted norm score from
the inverse regression model per group. This helps to inspect the precision
of the modeling process. The scores should not deviate too far from
regression line. The computation of the standard error is based on Oosterhuis, van der
Ark and Sijtsma (2016).
}
\seealso{
plotNorm
}
