% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.raw.cnorm}
\alias{plot.raw.cnorm}
\title{S3 method for plotting raw scores against fitted raw scores of cnorm objects}
\usage{
plot.raw.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the raw data against the fitted scores from
the regression model per group. This helps to inspect the precision
of the modeling process. The scores should not deviate too far from
regression line.
}
\seealso{
plotRaw

Other plot: 
\code{\link{plot.curves.cnorm}()},
\code{\link{plot.density.cnorm}()},
\code{\link{plot.derivative.cnorm}()},
\code{\link{plot.norm.cnorm}()},
\code{\link{plot.series.cnorm}()},
\code{\link{plot.subset.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNormCurves}()},
\code{\link{plotNorm}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
