% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.cnorm}
\alias{plot.cnorm}
\title{S3 method for plotting percentiles of cnorm objects}
\usage{
plot.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the norm curves based on the regression model against
the actual percentiles from the raw data.
}
\seealso{
plotPercentiles

Other plot: 
\code{\link{plot.curves.cnorm}()},
\code{\link{plot.density.cnorm}()},
\code{\link{plot.derivative.cnorm}()},
\code{\link{plot.norm.cnorm}()},
\code{\link{plot.raw.cnorm}()},
\code{\link{plot.series.cnorm}()},
\code{\link{plot.subset.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNormCurves}()},
\code{\link{plotNorm}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
