% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.series.cnorm}
\alias{plot.series.cnorm}
\title{S3 method for plotting a series of percentile curves of cnorm objects}
\usage{
plot.series.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
This functions makes use of 'plotPercentiles' to generate a series of plots
with different number of predictors. It draws on the information provided by the model object
to determine the bounds of the modeling (age and standard score range). It can be used as an
additional model check to determine the best fitting model.
}
\seealso{
plotPercentileSeries

Other plot: 
\code{\link{plot.curves.cnorm}()},
\code{\link{plot.density.cnorm}()},
\code{\link{plot.derivative.cnorm}()},
\code{\link{plot.norm.cnorm}()},
\code{\link{plot.percentiles.cnorm}()},
\code{\link{plot.raw.cnorm}()},
\code{\link{plot.subset.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNormCurves}()},
\code{\link{plotNorm}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
