% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.subset.cnorm}
\alias{plot.subset.cnorm}
\title{S3 method for plotting information function of continuous norming model of cnorm objects}
\usage{
plot.subset.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
Plots the information criterion - either Cp (default) or BIC - against
the adjusted R square of the feature selection in the modeling process.
Both BIC and Mallow's Cp are measures to avoid over-fitting. Please
choose the model that has a high information criterion, while modeling
the original data as close as possible.
}
\seealso{
plotSubset
}
