% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plotCnorm}
\alias{plotCnorm}
\alias{plot.cnorm}
\title{General convenience plotting function}
\usage{
plotCnorm(x, y, ...)

\method{plot}{cnorm}(x, y = "percentiles", ...)
}
\arguments{
\item{x}{a cnorm object}

\item{y}{The type of plot as a string or index: 'raw' (1), 'norm' (2),
'curves' (3), 'percentiles' (4), 'density' (5), 'series' (6),
'subset' (7) or 'derivative' (8). Defaults to 'percentiles'.}

\item{...}{additional parameters for the specific plotting function}
}
\description{
General convenience plotting function
}
