% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPercentiles}
\alias{plotPercentiles}
\title{Plot norm curves against actual percentiles}
\usage{
plotPercentiles(
  model,
  minRaw = NULL,
  maxRaw = NULL,
  minAge = NULL,
  maxAge = NULL,
  raw = NULL,
  group = NULL,
  percentiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  scale = NULL,
  title = NULL,
  subtitle = NULL,
  points = FALSE
)
}
\arguments{
\item{model}{The Taylor polynomial regression model object from the cNORM}

\item{minRaw}{Lower bound of the raw score (default = 0)}

\item{maxRaw}{Upper bound of the raw score}

\item{minAge}{Variable to restrict the lower bound of the plot to a specific age}

\item{maxAge}{Variable to restrict the upper bound of the plot to a specific age}

\item{raw}{The name of the raw variable}

\item{group}{The name of the grouping variable; the distinct groups are automatically
determined}

\item{percentiles}{Vector with percentile scores, ranging from 0 to 1 (exclusive)}

\item{scale}{The norm scale, either 'T', 'IQ', 'z', 'percentile' or
self defined with a double vector with the mean and standard deviation}

\item{title}{custom title for plot}

\item{subtitle}{custom title for plot}

\item{points}{Logical indicating whether to plot the data points. Default is TRUE.}
}
\description{
The function plots the norm curves based on the regression model against
the actual percentiles from the raw data. Applicable only for Taylor polynomial models.
For parametric models (Beta-Binomial, CMP, or SHASH), please use \code{plot(model, age, score)} instead.
}
\seealso{
plotNormCurves, plotPercentileSeries

Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnorm]{plot.cnorm()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
