% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppvt}
\alias{ppvt}
\title{Vocabulary development from 4 to 16}
\format{A data frame with 5600 rows and 9 variables:
\describe{
  \item{age}{the chronological age of the child}
  \item{group}{the according age group, e.g. age group 4 consists of children age 3.5 to 4.5}
  \item{sex}{the sex of the test taker, 1=male, 2=female}
  \item{schoolform}{the type of school in the German school system, the child is in. 0=missing, 1=preschool,
  2=elementary school, 3=Gesamtschule, 4=secondary modern school (Hauptschule), 5 = senior high school
  (Realschule), 6 = grammar school (Gymnasium), 7 = Training on the job (Berufsschule), 8 = university,
  9 = special educational services}
  \item{grade}{the grade, in case the child is in the school}
  \item{migration}{the migration backgroun with 0 = no migration, 1 = either one of the parents or the child immigrated}
  \item{language}{language spoken at home; 0 = missing, 1 = German, 2 = Turkish, 3 = Russian, 4 = Polish,
  5 = Spanish, 6 = Italian, 7 = other}
  \item{raw}{the raw score of the student, spanning values from 0 to 228}
  \item{duration}{the time needed for completing the test in ms}
}}
\source{
\url{https://www.psychometrica.de/ppvt4.html}
}
\usage{
ppvt
}
\description{
A dataset based on an unstratified sample of PPVT4 data (German adaption), including demografic data like
age, sex, school form, migration background and language spoken at home. The PPVT4 consists of blocks of items with
12 items each. Each item consists of 4 pictures. The test taker is given a word orally and he or she has to point out
the picture matching the oral word. Bottom and ceiling blocks of items are determined according to age and performance. For
instance, when a student knows less than 4 word from a block of 12 items, the testing stops.
}
\references{
Lenhard, A., Lenhard, W., Segerer, R. & Suggate, S. (2015). Peabody Picture Vocabulary Test - Revision IV (Deutsche Adaption). Frankfurt a. M./Germany: Pearson Assessment.
}
\keyword{datasets}
