% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{predict.cnormBetaBinomial2}
\alias{predict.cnormBetaBinomial2}
\title{Predict Norm Scores from Raw Scores}
\usage{
\method{predict}{cnormBetaBinomial2}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class 'cnormBetaBinomial' or
'cnormBetaBinomial2'.}

\item{...}{Additional arguments passed to the prediction method:
\itemize{
   \item age A numeric vector of ages, same length as score.
   \item score A numeric vector of raw scores.
   \item range The range of the norm scores in standard deviations.
     Default is 3.
 }}
}
\value{
A numeric vector of norm scores.
}
\description{
This function calculates norm scores based on raw scores, age, and a
fitted cnormBetaBinomial2 model. See
\code{\link{predict.cnormBetaBinomial}} for details.
}
\seealso{
Other predict:
\code{\link[=derivationTable]{derivationTable()}},
\code{\link[=getNormCurve]{getNormCurve()}},
\code{\link[=normTable]{normTable()}},
\code{\link[=predict.cnormBetaBinomial]{predict.cnormBetaBinomial()}},
\code{\link[=predict.cnormCMP]{predict.cnormCMP()}},
\code{\link[=predict.cnormShash]{predict.cnormShash()}},
\code{\link[=predictNorm]{predictNorm()}},
\code{\link[=predictRaw]{predictRaw()}},
\code{\link[=rawTable]{rawTable()}}
}
\concept{predict}
