% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{predict.norm.cnorm}
\alias{predict.norm.cnorm}
\title{S3 method for predicting single norm score for a given raw score and age}
\usage{
predict.norm.cnorm(norm, age, object, ...)
}
\arguments{
\item{norm}{A specific raw score}

\item{age}{The age within the continuous model}

\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
This function conducts this reverse
transformation via a numerical solution: A precise norm table is generated and
the closest fitting norm score for a raw score is returned.
}
\seealso{
predictNorm

Other predict: 
\code{\link{derivationTable}()},
\code{\link{getNormCurve}()},
\code{\link{normTable}()},
\code{\link{predict.raw.cnorm}()},
\code{\link{predictNorm}()},
\code{\link{predictRaw}()},
\code{\link{rawTable}()},
\code{\link{table.cnorm}()},
\code{\link{table.raw.cnorm}()}
}
\concept{predict}
