% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{predict.raw.cnorm}
\alias{predict.raw.cnorm}
\title{S3 method for predicting single raw score for a given norm score and age}
\usage{
predict.raw.cnorm(raw, age, object, ...)
}
\description{
Most elementary function to predict raw score based on Location (L, T score),
Age (grouping variable) and the coefficients from a regression model.
}
\seealso{
predictRaw
}
