% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{predictNorm.betabinomial}
\alias{predictNorm.betabinomial}
\title{Predict Norm Scores from Raw Scores}
\usage{
predictNorm.betabinomial(raw, age, model, n = NULL, range = 3)
}
\arguments{
\item{raw}{A numeric vector of raw scores.}

\item{age}{A numeric vector of ages, same length as raw.}

\item{model}{A fitted model object of class "cnormBetaBinomial".}

\item{n}{The number of items resp. the maximum score.}

\item{range}{The range of the norm scores in standard deviations. Default is 3. Thus, scores in the
range of +/- 3 standard deviations are considered.}
}
\value{
A numeric vector of norm scores.
}
\description{
This function calculates norm scores based on raw scores, age, and a fitted cnormBetaBinomial model.
}
\details{
The function first predicts the alpha and beta parameters of the beta-binomial distribution
for each age using the provided model. It then calculates the cumulative probability for
each raw score given these parameters. Finally, it converts these probabilities to the
norm scale specified in the model.
}
\examples{
# Assuming you have a fitted model named 'bb_model':
model <- cnorm.betabinomial(ppvt$age, ppvt$raw)
raw <- c(100, 121, 97, 180)
ages <- c(7, 8, 9, 10)
norm_scores <- predictNorm.betabinomial(raw, ages, model)

}
