% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNormValue}
\alias{predictNormValue}
\title{Retrieve norm value for raw score at a specific age}
\usage{
predictNormValue(raw, A, model, min = 25, max = 75, precision = 0.1)
}
\arguments{
\item{raw}{The raw value}

\item{A}{the age}

\item{model}{The regression model}

\item{min}{The lower bound of the norm value range}

\item{max}{The upper bound of the norm value range}

\item{precision}{The precision for the norm value generation with lower values
indicating a higher precision. In case of T values, precision = 0.1 is sufficient.}
}
\value{
The predicted norm value for a raw value
}
\description{
In real test scenarios, usually the results are available as raw values, for
which norm values have to be looked up. This function conducts this reverse
transformation via a numerical solution: A precise norm table is generated and
the closest fitting norm value for a raw value is returned.
}
\examples{
normData <- prepareData()
m <- bestModel(data=normData)

# return norm value for raw value 21 for grade 2, month 9
specificNormValue <- predictNormValue(21, 2.75, m)
}
