% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepareData}
\alias{prepareData}
\title{Set up example dataset and compute model}
\usage{
prepareData(data = NULL, group = "group", raw = "raw",
  age = "group")
}
\arguments{
\item{data}{data.frame with a grouping variable named 'group' and a raw score variable
named 'raw'. In case no object is provided, cNORM uses the inbuilt sample data to
demonstrate the procedure.}

\item{group}{grouping variable in the data, e. g. age groups, grades ...}

\item{raw}{the raw scores}

\item{age}{the continuous explanatory variable; by default set to "group"}
}
\value{
data frame including the norm scores, powers and interactions of the norm score and
grouping variable
}
\description{
This is a convenience method to either load the inbuilt sample dataset, or
to provide a data frame with the variables "raw" (for the raw scores) and "group"
The function ranks the data within groups, computes norm values, powers of the norm
scores and interactions. Afterwards the best fitting model is determined, based on
all default parameters.
}
\examples{
# conducts ranking and computation of powers and interactions with the 'elfe' dataset
data.elfe <- prepareData()

# variable names can be specified as well, here with the BMI data included in the package
data.bmi <- prepareData(CDC, group="group", raw="bmi", age="age")
}
