% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prettyPrint}
\alias{prettyPrint}
\title{Format raw and norm tables
The function takes a raw or norm table, condenses intervals at the bottom and top
and round the numbers to meaningful interval.}
\usage{
prettyPrint(table)
}
\arguments{
\item{table}{The table to format}
}
\value{
formatted table
}
\description{
Format raw and norm tables
The function takes a raw or norm table, condenses intervals at the bottom and top
and round the numbers to meaningful interval.
}
