% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{print.cnormShash}
\alias{print.cnormShash}
\title{Print method for Sinh-Arcsinh objects}
\usage{
\method{print}{cnormShash}(x, ...)
}
\arguments{
\item{x}{A cnormShash object}

\item{...}{Additional arguments}
}
\description{
Print method for Sinh-Arcsinh objects
}
