% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.cnorm}
\alias{print.cnorm}
\title{S3 method for printing subset information of cnorm object}
\usage{
print.cnorm(object, ...)
}
\description{
After conducting the model fitting procedure on the data set, the best fitting
model has to be chosen. The print function shows the R2 and other information criteria
on the different best fitting models with increasing number of predictors.
}
\seealso{
printSubset
}
