% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{printExtrapolationWarning}
\alias{printExtrapolationWarning}
\title{Internal function to check for horizontal and vertical extrapolation}
\usage{
printExtrapolationWarning(model, minA, maxA, minL, maxL)
}
\arguments{
\item{model}{The regression model}

\item{minA}{The lower age bound}

\item{maxA}{The upper age bound}

\item{minL}{The lower norm value bound}

\item{maxL}{The upper norm value bound}
}
\description{
Regression model only work in a specific range and extrapolation horizontally (outside
the original range) or vertically (extreme norm values) might lead to inconsistent
results. The function prints messages, indicating extrapolation.
}
