% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{pshash}
\alias{pshash}
\title{Sinh-Arcsinh (ShaSh) probability}
\usage{
pshash(
  q,
  mu = 0,
  sigma = 1,
  epsilon = 0,
  delta = 1,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter (> 0)}

\item{epsilon}{skewness parameter}

\item{delta}{tail weight parameter (> 0)}

\item{lower.tail}{logical; if TRUE, probabilities are P[X <= x]}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Density, distribution function, quantile function and random generation
for the Sinh-Arcsinh distribution
}
\keyword{internal}
