% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{rangeCheck}
\alias{rangeCheck}
\title{Check for horizontal and vertical extrapolation}
\usage{
rangeCheck(
  object,
  minAge = NULL,
  maxAge = NULL,
  minNorm = NULL,
  maxNorm = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{The regression model or a cnorm object}

\item{minAge}{The lower age bound}

\item{maxAge}{The upper age bound}

\item{minNorm}{The lower norm value bound}

\item{maxNorm}{The upper norm value bound}

\item{digits}{The precision for rounding the norm and age data}

\item{...}{additional parameters}
}
\value{
the report
}
\description{
Regression models only work in a specific range and extrapolation
horizontally (outside the original range) or vertically (extreme norm
scores) might lead to inconsistent results. The function generates a
message, indicating extrapolation and the range of the original data.
}
\examples{
\dontrun{
  m <- cnorm(raw = elfe$raw, group = elfe$group)
  rangeCheck(m)
}
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=diagnostics]{diagnostics()}},
\code{\link[=modelSummary]{modelSummary()}},
\code{\link[=normTable]{normTable()}},
\code{\link[=printSubset]{printSubset()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=weightedAverageModel]{weightedAverageModel()}}
}
\concept{model}
