% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{rankByGroup}
\alias{rankByGroup}
\title{Determine the norm values of the participants in each subsample}
\usage{
rankByGroup(data, group = "group", method = 4, scale = "T")
}
\arguments{
\item{data}{data.frame with norm sample data}

\item{group}{the grouping variable, e. g. grade}

\item{method}{Ranking method in case of bindings, please provide an index, choosin from the following
methods: 1 = Blom (1958), 2 = Tukey (1949), 3 = Van der Warden (1952), 4 = Rankit (default), 5 = Levenbach (1953),
6 = Filliben (1975), 7 = Yu & Huang (2001)}

\item{scale}{type of norm scale, either T (default), IQ or z; you can provide a double vector with the mean and
standard deviation as well, f. e. c(10, 3) for Wechsler scale index points}
}
\value{
the dataset with the percentiles and norm scales per group
}
\description{
This is the initial step, usually done in all kinds of test norming projects,
after the scale is constructed and the norm sample is established. First,
the data is grouped according to a grouping variable and afterwards, the percentiles
for each raw value is retrieved. The percentile can be used for the modeling
procedure, but in case, the samples to not deviate too much from normality,
T, IQ or z values can be computed via a normal rank procedure based on the
inverse cumulative normal distribution. In case of bindings, we use the medium rank
and there are different methods for estimating the percentiles (default RankIt).
}
\examples{
#Transformation with default parameters: RandIt and converting to T values
normData <- rankByGroup(elfe, group = "group")

#Transformation into Wechsler points with Yu & Huang (2001) ranking procedure
normData <- rankByGroup(elfe, group = "group", method = 7, scale=c(10, 3))


}
