% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rawTableQuick}
\alias{rawTableQuick}
\title{Internal method for two step shotgun search of raw -> norm value table}
\usage{
rawTableQuick(A, model, minRaw = NULL, maxRaw = NULL, minNorm = NULL,
  maxNorm = NULL, step = 1, precision = 0.1, descend = FALSE)
}
\arguments{
\item{A}{the age}

\item{model}{The regression model}

\item{minRaw}{The lower bound of the raw score range}

\item{maxRaw}{The upper bound of the raw score range}

\item{minNorm}{Clipping parameter for the lower bound of norm scores (default 25)}

\item{maxNorm}{Clipping parameter for the upper bound of norm scores (default 75)}

\item{step}{Stepping parameter for the raw values (default 1)}

\item{precision}{Precision for the norm score estimation. Lower values indicate
higher precision (default .1)}

\item{descend}{Reverse raw value order. If set to TRUE, lower raw scores
indicate higher performance. Relevant f. e. in case of modelling errors}
}
\value{
data.frame with raw scores and the predicted norm scores
}
\description{
This function is comparable to 'rawTable', but uses an optimization to reduce computational
resources. It is the default method in the 'rawTable' function. TODO: It still has to be checked for
descending values.
}
