% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{rshash}
\alias{rshash}
\title{Sinh-Arcsinh (shash) random generation}
\usage{
rshash(n, mu = 0, sigma = 1, epsilon = 0, delta = 1)
}
\arguments{
\item{mu}{location parameter}

\item{sigma}{scale parameter (> 0)}

\item{epsilon}{skewness parameter}

\item{delta}{tail weight parameter (> 0)}
}
\description{
Density, distribution function, quantile function and random generation
for the Sinh-Arcsinh distribution
}
\keyword{internal}
