% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simData}
\alias{simData}
\title{Simulate curve linear and skewed sample data}
\usage{
simData(n, minAge = 1, maxAge = 4, minRaw = 0, maxRaw = 50,
  nGroups = 4, pow = 1, slant = 0, k = 4)
}
\arguments{
\item{n}{number of observations}

\item{minAge}{the lower age bound}

\item{maxAge}{the upper age bound}

\item{minRaw}{the lower bound of raw values}

\item{maxRaw}{the upper bound of raw values}

\item{nGroups}{number of samples, that should be drawn}

\item{pow}{Power factor. Values < 1 represent root functions (e. g. pow = .5)}

\item{slant}{Measure for skewness within each sample. Use 0 for symetric normal distributions,
values > 1 for positively skewed data (e. g. 10), values lower than 0 for negatively skewed
distributions. The according function is based on Azzalini (2014)}

\item{k}{The power paramerer for the computePowers and bestMode function}
}
\value{
The simulated sample data set with ranking and powers already applied
}
\description{
Simulate curve linear and skewed sample data
}
\examples{
sample <- simData(100, minAge = 1, maxAge = 4, nGroups = 4)
model <- bestModel(sample)
plotPercentiles(sample, model)
}
\references{
Azzalini, A. with the collaboration of Capitanio, A. (2014). The Skew-Normal and Related Families. Cambridge University Press, IMS Monographs series.
}
