% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simMean}
\alias{simMean}
\title{Simulate mean per age}
\usage{
simMean(age)
}
\arguments{
\item{age}{the age variable}
}
\value{
return predicted means
}
\description{
Simulate mean per age
}
\examples{
\dontrun{
x <- simMean(a)
}
}
