% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simSD}
\alias{simSD}
\title{Simulate sd per age}
\usage{
simSD(age)
}
\arguments{
\item{age}{the age variable}
}
\value{
return predicted sd
}
\description{
Simulate sd per age
}
\examples{
\dontrun{
x <- simSD(a)
}
}
