% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simValue}
\alias{simValue}
\title{Simulate value per age based on regression coefficients}
\usage{
simValue(age, coefficients = c(0, 1.5, -0.05, 0, 1e-04))
}
\arguments{
\item{age}{the age variable}

\item{coefficients}{the coefficients of the regression function with increasing power
(first value intercept, second * age^1 ...)}
}
\value{
return predicted value
}
\description{
Simulate value per age based on regression coefficients
}
\examples{
\dontrun{
x <- simValue(a, coefficients = c(0, 1.5, -0.05, 0, 0.0001))
}
}
