% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{skewedNormal}
\alias{skewedNormal}
\title{Internal method for creating a skewed normal distribution}
\usage{
skewedNormal(n, location, scale, slant = 10)
}
\arguments{
\item{n}{number of observations}

\item{location}{mean of the distribution}

\item{scale}{the standard deviation}

\item{slant}{Measure for skewness}
}
\value{
a skewed normal distribution
}
\description{
The function generates a skewed normal distribution based on the SN
package by Adelchi Azzalini (http://azzalini.stat.unipd.it/SN/). It
is used for internal testing purposes.
}
\references{
Azzalini, A. with the collaboration of Capitanio, A. (2014).
The Skew-Normal and Related Families. Cambridge University Press,
IMS Monographs series.
}
