% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardize}
\alias{standardize}
\title{Standardize a numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{A numeric vector to be standardized.}
}
\value{
A numeric vector of the same length as x, containing the standardized values.
}
\description{
This function standardizes a numeric vector by subtracting the mean
and dividing by the standard deviation. The resulting vector will have
a mean of 0 and a standard deviation of 1.
}
\examples{
data <- c(1, 2, 3, 4, 5)
standardized_data <- standardize(data)
print(standardized_data)

}
