% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking.R
\name{standardizeMargins}
\alias{standardizeMargins}
\title{Function for standardizing population margins with respect to the sample size.
This step is necessary, since the in advance of the computation it must be
guaranteed that all marginals are passed to the rake function of survey
with respect to the same sample size, i.e., that all marginals are passed
to the rake function in realtion to the norm sample size.
to the rake function in reaction to the norm sample size.}
\usage{
standardizeMargins(data, population.margins)
}
\arguments{
\item{data}{data.frame with norm sample data.}

\item{population.margins}{List of data frames containing the marginals of
the specified stratification variables. The first column of the data frame
must contain the different levels of the variable and the second rows the
frequency.}
}
\description{
Function for standardizing population margins with respect to the sample size.
This step is necessary, since the in advance of the computation it must be
guaranteed that all marginals are passed to the rake function of survey
with respect to the same sample size, i.e., that all marginals are passed
to the rake function in realtion to the norm sample size.
to the rake function in reaction to the norm sample size.
}
