% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.cnorm}
\alias{summary.cnorm}
\title{S3 method for printing the results and regression function of a cnorm model}
\usage{
\method{summary}{cnorm}(object, ...)
}
\arguments{
\item{object}{A regression model or cnorm object}

\item{...}{additional parameters}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
S3 method for printing the results and regression function of a cnorm model
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=modelSummary]{modelSummary()}},
\code{\link[=print.cnorm]{print.cnorm()}},
\code{\link[=printSubset]{printSubset()}},
\code{\link[=rangeCheck]{rangeCheck()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=weightedAverageModel]{weightedAverageModel()}}
}
\concept{model}
