% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{summary.cnormBetaBinomial2}
\alias{summary.cnormBetaBinomial2}
\title{Summarize a Beta-Binomial Continuous Norming Model}
\usage{
\method{summary}{cnormBetaBinomial2}(object, ...)
}
\arguments{
\item{object}{An object of class "cnormBetaBinomial" or
"cnormBetaBinomial2", typically the result of a call to
\code{\link{cnorm.betabinomial}}.}

\item{...}{Additional arguments passed to the summary method:
\itemize{
   \item age An optional numeric vector of age values.
   \item score An optional numeric vector of raw scores.
   \item weights An optional numeric vector of weights.
 }}
}
\value{
Invisibly returns a list containing detailed diagnostic
  information about the model.
}
\description{
Summary method for objects of class "cnormBetaBinomial2". See
\code{\link{summary.cnormBetaBinomial}} for details.
}
\seealso{
\code{\link{cnorm.betabinomial}}, \code{\link{diagnostics.betabinomial}}
}
