% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.cnorm}
\alias{summary.cnorm}
\title{S3 method for printing model summary of cnorm object}
\usage{
summary.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
S3 method for printing model summary of cnorm object
}
\seealso{
Other model: 
\code{\link{bestModel}()},
\code{\link{check.cnorm}()},
\code{\link{checkConsistency}()},
\code{\link{cnorm.cv}()},
\code{\link{cv.cnorm}()},
\code{\link{derive}()},
\code{\link{modelSummary}()},
\code{\link{print.cnorm}()},
\code{\link{printSubset}()},
\code{\link{rangeCheck}()},
\code{\link{regressionFunction}()}
}
\concept{model}
