% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{table.raw.cnorm}
\alias{table.raw.cnorm}
\title{S3 method for creating a raw score table for specific age(s) on the basis of a cnorm object}
\usage{
table.raw.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
This function generates a norm table for raw scores at specific age(s) based on the regression
model by assigning raw scores to norm scores.
}
\seealso{
rawTable

Other predict: 
\code{\link{derivationTable}()},
\code{\link{getNormCurve}()},
\code{\link{normTable}()},
\code{\link{predict.norm.cnorm}()},
\code{\link{predict.raw.cnorm}()},
\code{\link{predictNorm}()},
\code{\link{predictRaw}()},
\code{\link{rawTable}()},
\code{\link{table.cnorm}()}
}
\concept{predict}
