% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{table.cnorm}
\alias{table.cnorm}
\title{S3 method for creating a norm table for specific age(s) on the basis of a cnorm object}
\usage{
table.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
This function generates a norm table for specific age(s) based on the regression
model by assigning raw scores to norm scores.
}
\seealso{
normTable
}
