% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{weighted.quantile.type7}
\alias{weighted.quantile.type7}
\title{Weighted type7 quantile estimator}
\usage{
weighted.quantile.type7(x, probs, weights = NA)
}
\arguments{
\item{x}{A numerical vector}

\item{probs}{Numerical vector of quantiles}

\item{weights}{A numerical vector with weights; should have the same length as x}
}
\value{
the quantiles
}
\description{
Computes weightes quantiles; code from Andrey Akinshin via https://aakinshin.net/posts/weighted-quantiles/
Code made available via the CC BY-NC-SA 4.0 license
}
