% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{bestModel}
\alias{bestModel}
\title{Retrieve the best fitting regression model based on powers of A, L and interactions}
\usage{
bestModel(data, R2 = 0.99, k = 4, terms = 0)
}
\arguments{
\item{data}{The preprocessed dataset, which should include the variables 'raw'
and the powers and interactions of the norm value (L = Location; usually T values)
and an explanatory variably (usually age = A)}

\item{R2}{Adjusted R square as a stopping criterion for the model building
(default R2 = 0.99)}

\item{k}{The power constant. Higher values result in more detailed approximations
but have the danger of over-fit (default = 4, max = 6)}

\item{terms}{Selection criterion for model building. The best fitting model with
this number of terms is used}
}
\value{
The model meeting the R2 criteria with coefficients and variable selection
in model$coefficients. Use \code{plotSubset(model)} and
\code{plotPercentiles(data, model)} to inspect model
}
\description{
The function computes a series of regressions with an increasing number of predictors and
takes the best fitting model per step. The aim is to find a model with as few predictors
as possible, which at the same time manages to explain as much variance as possible from
the original data. In psychometric test construction, this approach can be used to smooth
the data and eliminate noise from norm sample stratification, while preserving the overall
diagnostic information. Values around R2 = .99 usually show excellent results. The selection
of the model can either be based on the number of terms in the regression functions or the
share of explained variance of the model (R2). If both are specified, first the method tries
to select the model based on the number of terms and in case, this does not work, use R2
instead. Pushing R2 by setting the number of terms, the R2 cut off and k to high values
might lead to on over-fit, so be careful! These paramters depend on the distribution of
the norm data. As a rule of thumb, terms = 5 or R2 = .99 and k = 4 is a good starting point
for the analyses.
\code{plotSubset(model)} can be used to weigh up R2 and information criteria (Cp, an AIC like measure)
and predicted versus actual values can be plotted with 'plotValues' and 'plotPercentiles'.
}
\examples{
normData <- prepareData()
model <- bestModel(normData)
plotSubset(model)
plotPercentiles(normData, model)
}
