% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{buildFunction}
\alias{buildFunction}
\title{Build regression function for bestModel}
\usage{
buildFunction(raw, k, t, age)
}
\arguments{
\item{raw}{name of the raw score variable}

\item{k}{the power degree for location}

\item{t}{the power degree for age}

\item{age}{use age}
}
\value{
regression function
}
\description{
Build regression function for bestModel
}
