% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{cnorm.betabinomial}
\alias{cnorm.betabinomial}
\title{Fit a heteroscedastic regression model}
\usage{
cnorm.betabinomial(
  time,
  score,
  mu = 3,
  sigma = 3,
  n = NULL,
  control = NULL,
  scale = "T",
  descend = FALSE
)
}
\arguments{
\item{time}{A numeric vector of predictor values (e.g., age or time).}

\item{score}{A numeric vector of response values.}

\item{mu}{Integer specifying the degree of the polynomial for the mean model. Default is 2.}

\item{sigma}{Integer specifying the degree of the polynomial for the standard deviation model. Default is 1.}

\item{n}{Number of items in the test, resp. maximum score to be achieved}

\item{control}{A list of control parameters to be passed to the `optim` function.
If NULL, default values are used.}

\item{scale}{type of norm scale, either T (default), IQ, z or percentile (= no
transformation); a double vector with the mean and standard deviation can as
well, be provided f. e. c(10, 3) for Wechsler scale index points}

\item{descend}{ranking order (default descent = FALSE): inverses the
ranking order with higher raw scores getting lower norm scores; relevant
for example when norming error scores, where lower scores mean higher
performance}
}
\value{
A list of class "hetero_model" containing:
  \item{beta_est}{Estimated coefficients for the mean model}
  \item{gamma_est}{Estimated coefficients for the log-standard deviation model}
  \item{se}{Standard errors of the estimated coefficients}
  \item{mu}{Degree of the polynomial for the mean model}
  \item{sigma}{Degree of the polynomial for the standard deviation model}
  \item{result}{Full result from the optimization procedure}
}
\description{
This function fits a heteroscedastic regression model where both the mean and
standard deviation of the response variable are modeled as polynomial functions
of the predictor variable.
}
\details{
The function standardizes the input variables, fits polynomial models for both
the mean and standard deviation, and uses maximum likelihood estimation to
find the optimal parameters. The optimization is performed using the BFGS method.
}
\examples{
# Fit a heteroscedastic regression model to the PPVT data
model <- cnorm.betabinomial(ppvt$age, ppvt$raw)
summary(model)


}
