% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{cnorm.cv}
\alias{cnorm.cv}
\title{Cross validation for term selection}
\usage{
cnorm.cv(data, repetitions = 1, norms = TRUE, max = 12)
}
\arguments{
\item{repetitions}{number of repetitions for cross validation}

\item{norms}{determine norm score crossfit and R2 (if set to TRUE). The option is
computationally intensive and duration increases with sample size, number of
repetitions and maximum number of terms (max option).}

\item{max}{Maximum number of terms in model up to 2*k + k^2}

\item{d}{data.frame of norm sample with ranking, powers and interaction of L and A}
}
\description{
This function cross validates the model selection by drawing 90% of the data as
training data and 10% as the validation data. The cases are stratified by norm
group. Successive models are retrieved with increasing number of terms and the
RMSE is plotted for the training, validation and the complete dataset. Additionally
to this analysis on the raw score level, it is possible to estimate the mean norm score
reliability and crossfit measures. For this, please set the norms parameter to TRUE. Due
to the high computational load when computing norm scores, it takes time to finish
when doing repeated cv or comparing models up to the maximum number of terms.
}
\examples{
# plot cross validation RMSE by number of terms up to 9 with three repetitions
data <- prepareData()
cnorm.cv(data, 3, max=9, norms=FALSE)
}
