% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNORM.R
\name{cNORM.GUI}
\alias{cNORM.GUI}
\title{Launcher for the graphical user interface of cNORM for distribution free
continuous norming}
\usage{
cNORM.GUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Default TRUE; automatically open browser for GUI}
}
\description{
Launcher for the graphical user interface of cNORM for distribution free
continuous norming
}
\examples{
\dontrun{
# Launch graphical user interface
cNORM.GUI()
}
}
