% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking.R
\name{computeWeights}
\alias{computeWeights}
\title{Integration of raking into cNORM}
\usage{
computeWeights(data, population.margins)
}
\arguments{
\item{data}{data.frame with norm sample data.
The function is used to compute the actual raking weights using the survey package.}

\item{population.margins}{List of data frames containing the marginals of
the specified stratification variables. The first column of each data frame
must contain the different levels of one variable and the second rows the
frequency. Moreover, the name of the first column must match the corresponding
variable name, while the name of the second column must be "Freq". That means that
the marginals must be passed to function as a list of single data frames, one data frame
per stratification variable.}
}
\description{
Computes and standardize raking weights based on the implementation in
the survey R package.
Generation of weights to compensate for non-stratified samples
}
\details{
This function computes standardized raking weights based on the implementation in
the survey R package to overcome biases in norm samples. It generates weights, by drawing
on the information of population shares (e. g. for sex, ethnic group, region ...) and
subsequently reduces the influence of over-represented groups or increases underrepresented
cases. The returned weights or scaled to be larger than 0.
}
