% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{derivationTable}
\alias{derivationTable}
\title{Create a table based on first derivation of the regression model for specific age}
\usage{
derivationTable(A, model, min = 25, max = 75, step = 0.1)
}
\arguments{
\item{A}{the age}

\item{model}{The regression model}

\item{min}{The lower bound of the norm value range}

\item{max}{The upper bound of the norm value range}

\item{step}{Stepping parameter with lower values indicating higher precision}
}
\value{
data.frame with norm values and the predicted value based on the
derived regression function
}
\description{
In order to check model assumptions, a table of the first derivation of the model
coefficients is created.
TODO additional information necessary
}
\examples{
normData <- prepareData()
m <- bestModel(data=normData)
d <- derivationTable(6, m, step=0.5)
}
