% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{derive}
\alias{derive}
\title{First order derivative of regression model}
\usage{
derive(model)
}
\arguments{
\item{model}{The regression model}
}
\value{
The derived coefficients
}
\description{
Calculates the first order derivative of the location / norm value from the regression model. This
is useful for finding violations of model assumptions and problematic distribution features as
f. e. bottom and ceiling effects, non-progressive norm scores within an age group or in general
intersecting percentile curves.
}
\examples{
normData <- prepareData()
m <- bestModel(normData)
derivedCoefficients <- derive(m)
}
