% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\docType{data}
\name{elfe}
\alias{elfe}
\title{Sentence comprehension data from grade 2.0 to 5.0}
\format{A data frame with 1400 rows and 3 columns}
\usage{
data(elfe)
}
\description{
A dataset containing the raw data of 1400 students from grade 2 to 5
in the sentence comprehension test from ELFE 1-6 (Lenhard &
Schneider, 2006). In this test, students are presented lists of
sentences with one gap. The student has to fill in the correct solution
by selecting from a list of 5 alternatives per sentence. The text is speeded,
with a time cutoff of 180 seconds. The variables are as follows:
}
\details{
\itemize{
  \item personID. Identification number of each single student
  \item group. Grade level, beginning with 2.0 up to 5.0. Value of .0
  indicate the middle of the school year, 2.5 the end of the school year
  \item raw. The raw score of the test, ranging from 0 to 28
}
}
\keyword{datasets}
