% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{getNormScoreSE}
\alias{getNormScoreSE}
\title{Calculates the standard error (SE) of the norm scores}
\usage{
getNormScoreSE(model)
}
\arguments{
\item{model}{a cnorm object}
}
\value{
The standard error (SE) of the norm scores sensu Oosterhuis et al. (2016)
}
\description{
Calculates the standard error (SE) of the norm scores
}
\references{
Oosterhuis, H. E. M., van der Ark, L. A., & Sijtsma, K. (2016). Sample Size Requirements for Traditional and Regression-Based Norms. Assessment, 23(2), 191–202. https://doi.org/10.1177/1073191115580638
}
