% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the negative log-likelihood for a heteroscedastic regression model}
\usage{
log_likelihood(params, X, Z, y)
}
\arguments{
\item{params}{A numeric vector containing all model parameters. The first n_beta elements
are coefficients for the mean model, and the remaining elements are
coefficients for the log-standard deviation model.}

\item{X}{A matrix of predictors for the mean model.}

\item{Z}{A matrix of predictors for the log-standard deviation model.}

\item{y}{A numeric vector of response values.}
}
\value{
The negative log-likelihood of the model.
}
\description{
This function computes the negative log-likelihood for a heteroscedastic regression model
where both the mean and standard deviation are modeled as functions of predictors.
}
\keyword{internal}
