% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{modelSummary}
\alias{modelSummary}
\title{Prints the results and regression function of a cnorm model}
\usage{
modelSummary(object, ...)
}
\arguments{
\item{object}{A regression model or cnorm object}

\item{...}{additional parameters}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
Prints the results and regression function of a cnorm model
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=print.cnorm]{print.cnorm()}},
\code{\link[=printSubset]{printSubset()}},
\code{\link[=rangeCheck]{rangeCheck()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=summary.cnorm]{summary.cnorm()}}
}
\concept{model}
