% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{modelSummary}
\alias{modelSummary}
\title{Prints the results and regression function of a cnorm model}
\usage{
modelSummary(model)
}
\arguments{
\item{model}{A regression model or cnorm object}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
Prints the results and regression function of a cnorm model
}
\seealso{
Other model: 
\code{\link{bestModel}()},
\code{\link{check.cnorm}()},
\code{\link{checkConsistency}()},
\code{\link{cnorm.cv}()},
\code{\link{cv.cnorm}()},
\code{\link{derive}()},
\code{\link{print.cnorm}()},
\code{\link{printSubset}()},
\code{\link{rangeCheck}()},
\code{\link{regressionFunction}()},
\code{\link{summary.cnorm}()}
}
\concept{model}
