% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{modelSummary}
\alias{modelSummary}
\title{Prints the results and regression function of a cnorm model}
\usage{
modelSummary(object, ...)
}
\arguments{
\item{object}{A regression model or cnorm object}

\item{...}{additional parameters}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
Prints the results and regression function of a cnorm model
}
\seealso{
Other model: 
\code{\link{bestModel}()},
\code{\link{check.cnorm}()},
\code{\link{checkConsistency}()},
\code{\link{cnorm.cv}()},
\code{\link{cv.cnorm}()},
\code{\link{derive}()},
\code{\link{printSubset}()},
\code{\link{rangeCheck}()},
\code{\link{regressionFunction}()}
}
\concept{model}
