% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{normTable.betabinomial}
\alias{normTable.betabinomial}
\title{Calculate Cumulative Probabilities, Density, Percentiles, and Z-Scores for
Beta-Binomial Distribution}
\usage{
normTable.betabinomial(model, times, n = NULL, m = NULL)
}
\arguments{
\item{model}{The model, which was fitted using the `optimized.model` function.}

\item{times}{A numeric vector of time points at which to make predictions.}

\item{n}{The number of items resp. the maximum score.}

\item{m}{An optional stop criterion in table generation. Positive integer lower than n.}
}
\value{
A data frame with columns: x, Px, Pcum, Percentile, z, norm score
}
\description{
Calculate Cumulative Probabilities, Density, Percentiles, and Z-Scores for
Beta-Binomial Distribution
}
