% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{normTable}
\alias{normTable}
\title{Generate norm tables for all model types}
\usage{
normTable(A, model, ...)
}
\arguments{
\item{A}{Age / grouping value(s)}

\item{model}{A cnorm, cnormModel, cnormBetaBinomial, cnormShash or cnormCMP object}

\item{...}{Further arguments passed to the specific implementation}
}
\description{
Convenience function that selects the appropriate implementation based on
the class of \code{model}.
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=diagnostics]{diagnostics()}},
\code{\link[=modelSummary]{modelSummary()}},
\code{\link[=printSubset]{printSubset()}},
\code{\link[=rangeCheck]{rangeCheck()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=weightedAverageModel]{weightedAverageModel()}}
}
\concept{model}
