% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{normTable}
\alias{normTable}
\title{Create a norm table based on model for specific age}
\usage{
normTable(A, model, min = 25, max = 75, step = 0.1, descend = FALSE)
}
\arguments{
\item{A}{the age}

\item{model}{The regression model}

\item{min}{The lower bound of the norm value range}

\item{max}{The upper bound of the norm value range}

\item{step}{Stepping parameter with lower values indicating higher precision}

\item{descend}{Reverse raw value order. If set to TRUE, lower raw values
indicate higher performance. Relevent f. e. in case of modelling errors}
}
\value{
data.frame with norm values and the predicted raw value
}
\description{
This function generates a norm table for a specific age based on the regression
model by assigning raw values to norm values. Please specify the
range of norm values, you want to cover. A T value of 25 corresponds to a percentile
of .6. As a consequence, specifying a rang of T = 25 to T = 75 would cover 98.4 % of
the population. Please be careful when extrapolating vertically (at the lower and
upper end of the age specific distribution). Depending on the size of your norm
sample, extreme values with T < 20 or T > 80 might lead to inconsistent results.
}
\examples{
normData <- prepareData()
m <- bestModel(data=normData)
norms <- normTable(3.5, m, step=0.5)
}
