% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{plot.betabinomial}
\alias{plot.betabinomial}
\title{Plot cnormBetaBinomial Model with Data and Percentile Lines}
\usage{
\method{plot}{betabinomial}(
  model,
  age,
  score,
  weights = NULL,
  percentiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  points = TRUE
)
}
\arguments{
\item{model}{A fitted model object of class "cnormBetaBinomial" or "cnormBetaBinomial2".}

\item{age}{A vector the age data.}

\item{score}{A vector of the score data.}

\item{weights}{A vector of weights for each observation. Default is NULL (equal weights).}

\item{percentiles}{A vector with the percentiles to plot}

\item{points}{Logical indicating whether to plot the data points. Default is TRUE.}
}
\value{
A ggplot object.
}
\description{
This function creates a visualization of a fitted cnormBetaBinomial model,
including the original data points manifest percentiles and specified percentile lines.
}
\seealso{
Other plot: 
\code{\link{plot.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNorm}()},
\code{\link{plotNormCurves}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
