% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.cnorm}
\alias{plot.cnorm}
\title{S3 function for plotting cnorm objects}
\usage{
\method{plot}{cnorm}(x, y, ...)
}
\arguments{
\item{x}{the cnorm object}

\item{y}{the type of plot as a string, can be one of
'raw' (1), 'norm' (2), 'curves' (3), 'percentiles' (4), 'series' (5), 'subset' (6),
or 'derivative' (7), either as a string or the according index}

\item{...}{additional parameters for the specific plotting function}
}
\description{
S3 function for plotting cnorm objects
}
\seealso{
Other plot: 
\code{\link{plot.cnormBetaBinomial}()},
\code{\link{plot.cnormBetaBinomial2}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNorm}()},
\code{\link{plotNormCurves}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
