% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{plot.cnormBetaBinomial}
\alias{plot.cnormBetaBinomial}
\title{Plot cnormBetaBinomial Model with Data and Percentile Lines}
\usage{
\method{plot}{cnormBetaBinomial}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "cnormBetaBinomial" or
"cnormBetaBinomial2".}

\item{...}{Additional arguments passed to the plot method.
\itemize{
   \item age A vector of the age data.
   \item score A vector of the score data.
   \item weights An optional numeric vector of weights for each observation.
   \item percentiles An optional vector with the percentiles to plot.
   \item points Logical indicating whether to plot the data points.
     Default is TRUE.
   \item discrete Logical indicating whether to plot the discrete raw
     scores. Default is FALSE.
 }}
}
\value{
A ggplot object.
}
\description{
This function creates a visualization of a fitted cnormBetaBinomial model,
including the original data points, manifest percentiles and specified
percentile lines. Note that the beta-binomial model aims at discrete raw
scores. We decided to display continuous percentile lines nonetheless in
order to maintain visual comparability with other modelling techniques.
If you prefer discretization, set the "discrete" parameter to TRUE.
}
\seealso{
Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnorm]{plot.cnorm()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
