% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBinomial.R
\name{plot.cnormBetaBinomial2}
\alias{plot.cnormBetaBinomial2}
\title{Plot cnormBetaBinomial Model with Data and Percentile Lines}
\usage{
\method{plot}{cnormBetaBinomial2}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "cnormBetaBinomial" or "cnormBetaBinomial2".}

\item{...}{Additional arguments passed to the plot method.
\itemize{
   \item age A vector the age data.
   \item A vector of the score data.
   \item weights An optional numeric vector of weights for each observation.
   \item percentiles An optional vector with the percentiles to plot.
   \item points Logical indicating whether to plot the data points. Default is TRUE.
 }}
}
\value{
A ggplot object.
}
\description{
This function creates a visualization of a fitted cnormBetaBinomial model,
including the original data points manifest percentiles and specified percentile lines.
}
\seealso{
Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnorm]{plot.cnorm()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
