% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPCM.R
\name{plot.cnormCMP}
\alias{plot.cnormCMP}
\title{Plot CMP Model with Data and Percentile Lines}
\usage{
\method{plot}{cnormCMP}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "cnormCMP"}

\item{...}{Additional arguments including age, score, weights, percentiles, points}
}
\value{
A ggplot object
}
\description{
Plot CMP Model with Data and Percentile Lines
}
