% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{plot.cnormShaSh}
\alias{plot.cnormShaSh}
\title{Plot cnormShaSh Model with Data and Percentile Lines}
\usage{
\method{plot}{cnormShaSh}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "cnormShaSh"}

\item{...}{Additional arguments including age, score, weights, percentiles, points}
}
\value{
A ggplot object
}
\description{
Plot cnormShaSh Model with Data and Percentile Lines
}
