% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.curves.cnorm}
\alias{plot.curves.cnorm}
\title{S3 method for plotting percentile curves of cnorm objects}
\usage{
plot.curves.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the norm curves based on the regression model.
Please check the function for inconsistent curves: The different
curves should not intersect. Violations of this assumption are a strong
indication for violations of model assumptions in modeling the relationship between raw
and norm scores.
}
\seealso{
plotNormCurves
}
