% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.density.cnorm}
\alias{plot.density.cnorm}
\title{S3 method for plotting the density percentile curves of cnorm objects}
\usage{
plot.density.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the density  curves based on the regression model against
the actual percentiles from the raw data. As in 'plotNormCurves',
please check for inconsistent curves, especially curves showing implausible shapes as f. e.
violations of biuniqueness.
}
\seealso{
plotDensity

Other plot: 
\code{\link{plot.cnorm}()},
\code{\link{plot.curves.cnorm}()},
\code{\link{plot.derivative.cnorm}()},
\code{\link{plot.norm.cnorm}()},
\code{\link{plot.raw.cnorm}()},
\code{\link{plot.series.cnorm}()},
\code{\link{plot.subset.cnorm}()},
\code{\link{plotDensity}()},
\code{\link{plotDerivative}()},
\code{\link{plotNormCurves}()},
\code{\link{plotNorm}()},
\code{\link{plotPercentileSeries}()},
\code{\link{plotPercentiles}()},
\code{\link{plotRaw}()},
\code{\link{plotSubset}()}
}
\concept{plot}
