% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.cnorm}
\alias{plot.cnorm}
\title{S3 method for plotting percentiles of cnorm objects}
\usage{
plot.cnorm(object, ...)
}
\description{
S3 method for plotting percentiles of cnorm objects
}
\seealso{
plotPercentiles
}
