% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.cnorm}
\alias{plot.cnorm}
\title{S3 method for plotting percentiles of cnorm objects}
\usage{
plot.cnorm(object, ...)
}
\arguments{
\item{object}{cnorm object}

\item{...}{Additional parameters}
}
\description{
The function plots the norm curves based on the regression model against
the actual percentiles from the raw data.
}
\seealso{
plotPercentiles
}
