% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCnorm}
\alias{plotCnorm}
\title{General convenience plotting function}
\usage{
plotCnorm(x, y, ...)
}
\arguments{
\item{x}{a cnorm object}

\item{y}{the type of plot as a string or index}

\item{...}{additional parameters for the specific plotting function}
}
\description{
General convenience plotting function
}
