% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCnorm}
\alias{plotCnorm}
\title{General convenience plotting function}
\usage{
plotCnorm(x, y, ...)
}
\arguments{
\item{x}{a cnorm object}

\item{y}{the type of plot as a string, can be one of
'raw' (1), 'norm' (2), 'curves' (3), 'percentiles' (4), 'series' (5), 'subset' (6),
or 'derivative' (7), either as a string or the according index}

\item{...}{additional parameters for the specific plotting function}
}
\description{
General convenience plotting function
}
