% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCnorm}
\alias{plotCnorm}
\title{General convencience plotting function}
\usage{
plotCnorm(x, y, ...)
}
\arguments{
\item{x}{a cnorm object}

\item{y}{the type of plot as a string, can be one of
'raw', 'norm', 'curves', 'percentiles', 'series', 'subset', or 'derivative'}

\item{...}{additional parameters for the specific plotting function}
}
\description{
General convencience plotting function
}
