% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot the density function per group by raw score}
\usage{
plotDensity(
  model,
  minRaw = NULL,
  maxRaw = NULL,
  minNorm = NULL,
  maxNorm = NULL,
  group = NULL
)
}
\arguments{
\item{model}{The model from the bestModel function, a cnorm object, a cnormBetaBinomial, a cnormBetaBinomial2,
a cnormCMP, or a cnormShash object.}

\item{minRaw}{Lower bound of the raw score. If NULL, it's automatically determined based on the model type.}

\item{maxRaw}{Upper bound of the raw score. If NULL, it's automatically determined based on the model type.}

\item{minNorm}{Lower bound of the norm score. If NULL, it's automatically determined based on the model type.}

\item{maxNorm}{Upper bound of the norm score. If NULL, it's automatically determined based on the model type.}

\item{group}{Numeric vector specifying the age groups to plot. If NULL, groups are automatically selected.}
}
\value{
A ggplot object representing the density functions.
}
\description{
This function plots density curves based on the regression model against the raw scores.
It supports traditional continuous norming models, beta-binomial models, Conway-Maxwell-Poisson (CMP)
models, and SHASH models.
}
\details{
The function generates density curves (or PMFs for discrete models) for specified age groups,
allowing for easy comparison of score distributions across different ages.

For beta-binomial and CMP models, the display reflects the probability mass function, while for
continuous models (Taylor and SHASH), it displays the probability density function.
}
\seealso{
Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
