% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDerivative}
\alias{plotDerivative}
\title{Plot first order derivative of regression model}
\usage{
plotDerivative(
  model,
  minAge = NULL,
  maxAge = NULL,
  minNorm = NULL,
  maxNorm = NULL,
  stepAge = NULL,
  stepNorm = NULL,
  order = 1
)
}
\arguments{
\item{model}{The model from the bestModel function, a cnorm object.}

\item{minAge}{Minimum age to start checking.}

\item{maxAge}{Maximum age for checking.}

\item{minNorm}{Lower end of the norm score range.}

\item{maxNorm}{Upper end of the norm score range.}

\item{stepAge}{Stepping parameter for age.}

\item{stepNorm}{Stepping parameter for norm scores.}

\item{order}{Degree of the derivative (default = 1).}
}
\description{
This function plots the scores obtained via the first order derivative of the regression model.
Applicable only for Taylor polynomial models.
}
\seealso{
Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnorm]{plot.cnorm()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentileSeries]{plotPercentileSeries()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
