% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNorm}
\alias{plotNorm}
\title{Plot manifest and fitted norm scores}
\usage{
plotNorm(data, model, group = "", minNorm = NULL, maxNorm = NULL)
}
\arguments{
\item{data}{The raw data within a data.frame}

\item{model}{The regression model}

\item{group}{The grouping variable, use empty string "" for no group}

\item{minNorm}{lower bound of fitted norm scores}

\item{maxNorm}{upper bound of fitted norm scores}
}
\description{
The function plots the manifest norm score against the fitted norm score from
the inverse regression model per group. This helps to inspect the precision
of the modeling process. The scores should not deviate too far from
regression line.
}
\examples{
# Load example data set, compute model and plot results
\dontrun{
normData <- prepareData()
m <- bestModel(data = normData)
plotNorm(normData, m, group="group", minNorm=25, maxNorm=75)
}
}
