% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPercentileSeries}
\alias{plotPercentileSeries}
\title{Generates a series of plots with percentile curves for different models}
\usage{
plotPercentileSeries(
  model,
  start = 1,
  end = NULL,
  group = NULL,
  percentiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
  filename = NULL
)
}
\arguments{
\item{model}{The Taylor polynomial regression model object or a cnorm object}

\item{start}{Number of terms to start with (default 1)}

\item{end}{Number of terms to end with}

\item{group}{The name of the grouping variable}

\item{percentiles}{Vector with percentile scores}

\item{filename}{Prefix of the filename if saving to png}
}
\description{
This function makes use of 'plotPercentiles' to generate a series of plots
for models with an increasing number of terms.
}
\seealso{
plotPercentiles

Other plot:
\code{\link[=compare]{compare()}},
\code{\link[=plot.cnorm]{plot.cnorm()}},
\code{\link[=plot.cnormBetaBinomial]{plot.cnormBetaBinomial()}},
\code{\link[=plot.cnormBetaBinomial2]{plot.cnormBetaBinomial2()}},
\code{\link[=plotDensity]{plotDensity()}},
\code{\link[=plotDerivative]{plotDerivative()}},
\code{\link[=plotNorm]{plotNorm()}},
\code{\link[=plotNormCurves]{plotNormCurves()}},
\code{\link[=plotPercentiles]{plotPercentiles()}},
\code{\link[=plotRaw]{plotRaw()}},
\code{\link[=plotSubset]{plotSubset()}}
}
\concept{plot}
