% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotValues}
\alias{plotValues}
\title{Plot actual and predicted raw values}
\usage{
plotValues(data, model, group = "group")
}
\arguments{
\item{data}{The raw data within a data.frame}

\item{model}{The regression model}

\item{group}{The grouping variable}
}
\description{
The function plots the raw data against the predicted values from
the regression model per group. This helps to inspect the precision
of the modeling process. The values should not deviate too far from
regression line.
}
\examples{
# Load example data set, compute model and plot results
normData <- prepareData()
m <- bestModel(data = normData)
plotValues(normData, m, group="group")
}
