% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{predict.norm.cnorm}
\alias{predict.norm.cnorm}
\title{S3 method for predicting single norm score for a given raw score and age}
\usage{
predict.norm.cnorm(raw, age, object, ...)
}
\description{
This function conducts this reverse
transformation via a numerical solution: A precise norm table is generated and
the closest fitting norm score for a raw score is returned.
}
\seealso{
predictNorm
}
