% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{predictCoefficients_cmp}
\alias{predictCoefficients_cmp}
\title{Predict parameters for a CMP regression model}
\usage{
predictCoefficients_cmp(model, ages)
}
\arguments{
\item{model}{An object of class "cnormCMP"}

\item{ages}{A numeric vector of age points for prediction}
}
\value{
A data frame with predicted mu, nu and lambda = mu^nu (and optionally mean, variance, sd)
}
\description{
Predict parameters for a CMP regression model
}
\keyword{internal}
