% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{predictCoefficients_shash}
\alias{predictCoefficients_shash}
\title{Predict parameters for a ShaSh regression model}
\usage{
predictCoefficients_shash(model, ages)
}
\arguments{
\item{model}{An object of class "cnormShaSh"}

\item{ages}{A numeric vector of age points for prediction}
}
\value{
A data frame with predicted parameters and statistics
}
\description{
Predict parameters for a ShaSh regression model
}
\keyword{internal}
