% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepareData}
\alias{prepareData}
\title{Set up example dataset and compute model}
\usage{
prepareData()
}
\description{
This is a convenience method to load the inbuilt sample dataset, rank the
data and compute norm values, compute powers and interactions and
determine the best fitting model based on all default parameters.
}
\examples{
normData <- prepareData()
}
