% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{prepareData}
\alias{prepareData}
\title{Set up example dataset and compute model}
\usage{
prepareData(data = NULL, group = "group")
}
\arguments{
\item{data}{data.frame with a grouping variable named 'group' and a raw score variable
named 'raw'. In case no object is provided, cNORM uses the inbuilt sample data to demonstrate
the procedure}

\item{group}{manually specify the grouping variable in the data}
}
\value{
data frame including the norm scores, powers and interactions of the norm score and
grouping variable
}
\description{
This is a convenience method to either load the inbuilt sample dataset, or
to provide a data frame with the variables "raw" (for the raw scores) and "group"
The function ranks the data within groups, computes norm values, powers of the norm
scores and interactions. Afterwards the best fitting model is determined, based on
all default parameters.
}
\examples{
normData <- prepareData()
}
