% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPCM.R
\name{print.cnormCMP}
\alias{print.cnormCMP}
\title{Print method for CMP objects}
\usage{
\method{print}{cnormCMP}(x, ...)
}
\arguments{
\item{x}{A cnormCMP object}

\item{...}{Additional arguments}
}
\description{
Print method for CMP objects
}
