% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shash.R
\name{print.cnormShaSh}
\alias{print.cnormShaSh}
\title{Print method for cnormShaSh objects}
\usage{
\method{print}{cnormShaSh}(x, ...)
}
\arguments{
\item{x}{A cnormShaSh object}

\item{...}{Additional arguments}
}
\description{
Print method for cnormShaSh objects
}
