% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{printSubset}
\alias{printSubset}
\title{Print Model Selection Information}
\usage{
printSubset(x, ...)
}
\arguments{
\item{x}{Model output from 'bestModel' or a cnorm object.}

\item{...}{Additional parameters.}
}
\value{
Table with model information criteria.
}
\description{
Displays R^2 and other metrics for models with varying predictors, aiding in
choosing the best-fitting model after model fitting. The F-test compares each
model to the preceding (smaller) one, with degrees of freedom based on the
actual difference in the number of parameters.
}
\examples{
\dontrun{
  result <- cnorm(raw = elfe$raw, group = elfe$group)
  printSubset(result)
}
}
\seealso{
Other model:
\code{\link[=bestModel]{bestModel()}},
\code{\link[=checkConsistency]{checkConsistency()}},
\code{\link[=cnorm.cv]{cnorm.cv()}},
\code{\link[=derive]{derive()}},
\code{\link[=modelSummary]{modelSummary()}},
\code{\link[=print.cnorm]{print.cnorm()}},
\code{\link[=rangeCheck]{rangeCheck()}},
\code{\link[=regressionFunction]{regressionFunction()}},
\code{\link[=summary.cnorm]{summary.cnorm()}},
\code{\link[=weightedAverageModel]{weightedAverageModel()}}
}
\concept{model}
