% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{printSubset}
\alias{printSubset}
\title{Convenience method for printing model selection information}
\usage{
printSubset(model)
}
\arguments{
\item{model}{The model from the 'bestModel' function}
}
\value{
A table with information criteria
}
\description{
After conducting the model fitting procedure on the data set, the best fitting
model has to be chosen. The print function shows the R2 and other information
on the different best fitting models with increasing number of predictors.
}
\examples{
model <- bestModel(prepareData())
printSubset(model)
}
