% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{rangeCheck}
\alias{rangeCheck}
\title{Check for horizontal and vertical extrapolation}
\usage{
rangeCheck(model, minA = NULL, maxA = NULL, minL = NULL,
  maxL = NULL, digits = 3)
}
\arguments{
\item{model}{The regression model}

\item{minA}{The lower age bound}

\item{maxA}{The upper age bound}

\item{minL}{The lower norm value bound}

\item{maxL}{The upper norm value bound}

\item{digits}{The precision fr rounding the norm and age data}
}
\value{
the report
}
\description{
Regression model only work in a specific range and extrapolation horizontally (outside
the original range) or vertically (extreme norm values) might lead to inconsistent
results. The function generates a message, indicating extrapolation and the range of the original data.
}
\examples{
normData <- prepareData()
m <- bestModel(normData)
print(rangeCheck(m))
}
