% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rawTable}
\alias{rawTable}
\title{Create a table with norms assigned to raw values for a specific age based on the regression model}
\usage{
rawTable(A, model, min, max, minNorm = 25, maxNorm = 75, step = 1,
  precision = 0.01, descend = FALSE, quick = TRUE)
}
\arguments{
\item{A}{the age}

\item{model}{The regression model}

\item{min}{The lower bound of the raw value range}

\item{max}{The upper bound of the raw value range}

\item{minNorm}{Clipping parameter for the lower bound of norm values (default 25)}

\item{maxNorm}{Clipping parameter for the upper bound of norm values (default 25)}

\item{step}{Stepping parameter for the raw values (default 1)}

\item{precision}{Precision for the norm value estimation. Lower values indicate
higher precision (default .01)}

\item{descend}{Reverse raw value order. If set to TRUE, lower raw values
indicate higher performance. Relevant f. e. in case of modelling errors}

\item{quick}{Forces the use of a shotgun method to quickly find the norm values
with the desired precision. Reproduces the same results as the thorough search in
case the model assumptions are met.}
}
\value{
data.frame with raw values and the predicted norm value
}
\description{
This function is comparable to 'normTable', despite it reverses the assignment:
A table with raw values and the according norm values for a specific age based on the regression
model is generated. This way, the inverse function of the regression model is solved numerically with
brute force. Please specify the range of raw values, you want to cover. With higher precision
and smaller stepping, this function becomes computational intensive, especially when quick is set
to FALSE for a thorough search.
}
\examples{
# generate a norm table for the raw value range from 0 to 28 for month 7 of grade 3
normData <- prepareData()
m <- bestModel(data=normData)
table <- rawTable(3 + 7/12, m, 0, 28, precision=.01)
}
\seealso{
normTable
}
