% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{regressionFunction}
\alias{regressionFunction}
\title{Regression function}
\usage{
regressionFunction(model)
}
\arguments{
\item{model}{The regression model from the bestModel function}
}
\value{
The regression formula as a string
}
\description{
The method builds the regression function for the regression model,
including the beta weights.
It can be used to predict the raw values based on age and location.
}
\examples{
normData <- prepareData()
model <- bestModel(normData)
regressionFunction(model)
}
