% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{regressionFunction}
\alias{regressionFunction}
\title{Regression function}
\usage{
regressionFunction(model, raw = NULL, digits = NULL)
}
\arguments{
\item{model}{The regression model from the bestModel function or a cnorm object}

\item{raw}{The name of the raw value variable (default 'raw')}

\item{digits}{Number of digits for formatting the coefficients}
}
\value{
The regression formula as a string
}
\description{
The method builds the regression function for the regression model,
including the beta weights.
It can be used to predict the raw scores based on age and location.
}
\examples{
result <- cnorm(raw = elfe$raw, group = elfe$group)
regressionFunction(result)
}
