% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking.R
\name{standardizeRakingWeights}
\alias{standardizeRakingWeights}
\title{Function for standardizing raking weights}
\usage{
standardizeRakingWeights(weights)
}
\arguments{
\item{weights}{Raking weights computed by computeWeights()}
}
\value{
the standardized weights
}
\description{
Raking weights get divided by the smallest weight. Thereby, all weights
become larger or equal to 1 without changing the ratio of the weights
to each other.
}
\keyword{internal}
