% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.cnorm}
\alias{summary.cnorm}
\title{S3 method for printing the results and regression function of a cnorm model}
\usage{
\method{summary}{cnorm}(object, ...)
}
\arguments{
\item{object}{A regression model or cnorm object}

\item{...}{additional parameters}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
S3 method for printing the results and regression function of a cnorm model
}
\seealso{
Other model: 
\code{\link{bestModel}()},
\code{\link{checkConsistency}()},
\code{\link{cnorm.cv}()},
\code{\link{derive}()},
\code{\link{modelSummary}()},
\code{\link{print.cnorm}()},
\code{\link{printSubset}()},
\code{\link{rangeCheck}()},
\code{\link{regressionFunction}()}
}
\concept{model}
