% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{summary.cnorm}
\alias{summary.cnorm}
\title{Prints the results and regression function of a cnorm model}
\usage{
summary.cnorm(object, ...)
}
\arguments{
\item{model}{A regression model or cnorm object}
}
\value{
A report on the regression function, weights, R2 and RMSE
}
\description{
Prints the results and regression function of a cnorm model
}
