% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{weighted.quantile}
\alias{weighted.quantile}
\title{Weighted quantile estimator}
\usage{
weighted.quantile(x, probs, weights = NA, type = "Type7")
}
\arguments{
\item{x}{A numerical vector}

\item{probs}{Numerical vector of quantiles}

\item{weights}{A numerical vector with weights; should have the same length as x}

\item{type}{Type of estimator, can either be "Harrell-Davis" using a beta function to
approximate the weighted percentiles (Harrell & Davis, 1982) or "Type7" (default; Hyndman & Fan, 1996), an adaption
of the generic quantile function in R, including weighting. All code based on the work of Akinshin (2020).}
}
\value{
the weighted quantiles
}
\description{
Computes weighted quantiles; code from Andrey Akinshin via https://aakinshin.net/posts/weighted-quantiles/
Code made available via the CC BY-NC-SA 4.0 license
}
\references{
\enumerate{
  \item Harrell, F.E. & Davis, C.E. (1982). A new distribution-free quantile estimator. Biometrika, 69(3), 635-640.
  \item Hyndman, R. J. & Fan, Y. 1996. Sample quantiles in statistical packages, American Statistician 50, 361–365.
  \item Akinshin, A. (2020). Weighted quantile estimators. https://aakinshin.net/posts/weighted-quantiles/
}
}
