% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{weighted.rank}
\alias{weighted.rank}
\title{Weighted ranking}
\usage{
weighted.rank(x, weights = NULL, n = 1000)
}
\arguments{
\item{x}{A numerical vector}

\item{weights}{A numerical vector with weights; should have the same length as x}

\item{n}{Granularity for approximation}
}
\value{
weighted ranks
}
\description{
Conducts weighted ranking on the basis of the weighted Harrell-Davis quantile estimator
}
\references{
Harrell, F.E. and Davis, C.E., 1982. A new distribution-free quantile estimator. Biometrika, 69(3), pp.635-640.
}
