% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{weighted.rank}
\alias{weighted.rank}
\title{Weighted rank estimation}
\usage{
weighted.rank(x, weights = NULL)
}
\arguments{
\item{x}{A numerical vector}

\item{weights}{A numerical vector with weights; should have the same length as x}
}
\value{
the weighted absolute ranks
}
\description{
Conducts weighted ranking on the basis of sums of weights per unique raw score.
Please provide a vector with raw values and an additional vector with the weight of each
observation. In case the weight vector is NULL, a normal ranking is done. The vectors may not
include NAs and the weights should be positive non-zero values.
}
